+++ src/main/java/com/hbm/inventory/control_panel/Control.java
@@ -3,6 +3,8 @@ package com.hbm.inventory.control_panel;
 import com.hbm.inventory.control_panel.controls.ControlType;
 import com.hbm.inventory.control_panel.controls.configs.SubElementBaseConfig;
 import com.hbm.render.loader.IModelCustom;
+import it.unimi.dsi.fastutil.objects.Object2ObjectLinkedOpenHashMap;
+import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
 import net.minecraft.nbt.NBTBase;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.ResourceLocation;
@@ -16,22 +18,36 @@ import java.util.Map.Entry;
 
 public abstract class Control {
 
+	static final String PLACEHOLDER_META_TAG = "ntmPlaceholderMeta";
+	static final String PLACEHOLDER_BOX_MIN_X = "boxMinX";
+	static final String PLACEHOLDER_BOX_MIN_Y = "boxMinY";
+	static final String PLACEHOLDER_BOX_MAX_X = "boxMaxX";
+	static final String PLACEHOLDER_BOX_MAX_Y = "boxMaxY";
+	static final String PLACEHOLDER_SIZE_HEIGHT = "sizeHeight";
+	static final String PLACEHOLDER_HAS_BOUNDING_BOX = "hasBoundingBox";
+	static final String PLACEHOLDER_BOUNDS_MIN_X = "boundsMinX";
+	static final String PLACEHOLDER_BOUNDS_MIN_Y = "boundsMinY";
+	static final String PLACEHOLDER_BOUNDS_MIN_Z = "boundsMinZ";
+	static final String PLACEHOLDER_BOUNDS_MAX_X = "boundsMaxX";
+	static final String PLACEHOLDER_BOUNDS_MAX_Y = "boundsMaxY";
+	static final String PLACEHOLDER_BOUNDS_MAX_Z = "boundsMaxZ";
+
 	public String name;
 	public final String registryName;
 	public ControlPanel panel;
 	//Set of block positions this control is connected to. When an event is sent, it gets sent to each one
 	public List<BlockPos> connectedSet = new ArrayList<>();
 	//A map of event names to node system for events this control is sending out to connected blocks
-	public Map<String, NodeSystem> sendNodeMap = new HashMap<>();
+	public Map<String, NodeSystem> sendNodeMap = new Object2ObjectLinkedOpenHashMap<>();
 	//A map of event names to node systems for events this control is receiving
-	public Map<String, NodeSystem> receiveNodeMap = new HashMap<>();
+	public Map<String, NodeSystem> receiveNodeMap = new Object2ObjectLinkedOpenHashMap<>();
 	//A map of all variables, either used internally by the control or in the node systems
-	public Map<String, DataValue> vars = new HashMap<>();
-	public Map<String, DataValue> varsPrev = new HashMap<>();
+	public Map<String, DataValue> vars = new Object2ObjectLinkedOpenHashMap<>();
+	public Map<String, DataValue> varsPrev = new Object2ObjectLinkedOpenHashMap<>();
 	//A set of the custom variables the user is allowed to remove
-	public Set<String> customVarNames = new HashSet<>();
+	public Set<String> customVarNames = new ObjectOpenHashSet<>();
 	// map of (static) initial configurations for a control e.g. color, size
-	public Map<String, DataValue> configMap = new HashMap<>();
+	public Map<String, DataValue> configMap = new Object2ObjectLinkedOpenHashMap<>();
 	public float posX;
 	public float posY;
 
@@ -54,13 +70,13 @@ public abstract class Control {
 		return configMap;
 	}
 	public void applyConfigs(Map<String, DataValue> configs) {
-		configMap = configs;
+		configMap = new Object2ObjectLinkedOpenHashMap<>(configs);
 	}
 
 	public void renderBatched(){};
 	public void render(){};
 	public List<String> getOutEvents(){return Collections.emptyList();};
-	public List<String> getInEvents(){return Arrays.asList("tick", "initialize");};
+	public List<String> getInEvents(){return Arrays.asList("tick", "initialize", "redstone_input");};
 	@SideOnly(Side.CLIENT)
 	public abstract IModelCustom getModel();
 	@SideOnly(Side.CLIENT)
@@ -103,7 +119,7 @@ public abstract class Control {
 	}
 
 	public NBTTagCompound writeToNBT(NBTTagCompound tag){
-		tag.setString("name", ControlRegistry.getName(this.getClass()));
+		tag.setString("name", registryName);
 		tag.setString("myName", name);
 		NBTTagCompound vars = new NBTTagCompound();
 		for(Entry<String, DataValue> e : this.vars.entrySet()) {
@@ -148,23 +164,66 @@ public abstract class Control {
 			configs.setTag(e.getKey(), e.getValue().writeToNBT());
 		}
 		tag.setTag("configs", configs);
+		NBTTagCompound placeholderMeta = new NBTTagCompound();
+		float[] box = getBox();
+		if(box != null && box.length >= 4) {
+			placeholderMeta.setFloat(PLACEHOLDER_BOX_MIN_X, box[0]);
+			placeholderMeta.setFloat(PLACEHOLDER_BOX_MIN_Y, box[1]);
+			placeholderMeta.setFloat(PLACEHOLDER_BOX_MAX_X, box[2]);
+			placeholderMeta.setFloat(PLACEHOLDER_BOX_MAX_Y, box[3]);
+		}
+		float[] size = getSize();
+		if(size != null && size.length >= 3) {
+			placeholderMeta.setFloat(PLACEHOLDER_SIZE_HEIGHT, Math.max(size[2], 0.1F));
+		}
+		AxisAlignedBB boundingBox = getBoundingBox();
+		placeholderMeta.setBoolean(PLACEHOLDER_HAS_BOUNDING_BOX, boundingBox != null);
+		if(boundingBox != null) {
+			placeholderMeta.setDouble(PLACEHOLDER_BOUNDS_MIN_X, boundingBox.minX);
+			placeholderMeta.setDouble(PLACEHOLDER_BOUNDS_MIN_Y, boundingBox.minY);
+			placeholderMeta.setDouble(PLACEHOLDER_BOUNDS_MIN_Z, boundingBox.minZ);
+			placeholderMeta.setDouble(PLACEHOLDER_BOUNDS_MAX_X, boundingBox.maxX);
+			placeholderMeta.setDouble(PLACEHOLDER_BOUNDS_MAX_Y, boundingBox.maxY);
+			placeholderMeta.setDouble(PLACEHOLDER_BOUNDS_MAX_Z, boundingBox.maxZ);
+		}
+		tag.setTag(PLACEHOLDER_META_TAG, placeholderMeta);
 
 		return tag;
 	}
 	
 	public void readFromNBT(NBTTagCompound tag){
+		if(tag.hasKey("myName")) {
+			this.name = tag.getString("myName");
+		}
+
 		NBTTagCompound vars = tag.getCompoundTag("vars");
 		for(String k : vars.getKeySet()) {
 			NBTBase base = vars.getTag(k);
 			DataValue val = DataValue.newFromNBT(base);
-			if(val != null) {
-				this.vars.put(k, val);
+			if(val == null) {
+				throw new IllegalStateException("Failed to deserialize variable '" + k + "' for control '" + registryName + "'");
 			}
+			this.vars.put(k, val);
 		}
 		
 		sendNodeMap.clear();
 		receiveNodeMap.clear();
-		
+		customVarNames.clear();
+		connectedSet.clear();
+
+		NBTTagCompound customVarNames = tag.getCompoundTag("customvars");
+		for(int i = 0; i < customVarNames.getKeySet().size(); i ++){
+			this.customVarNames.add(customVarNames.getString("var"+i));
+		}
+
+		NBTTagCompound connectedSet = tag.getCompoundTag("conset");
+		for(int i = 0; i < connectedSet.getKeySet().size()/3; i ++){
+			int x = connectedSet.getInteger("px"+i);
+			int y = connectedSet.getInteger("py"+i);
+			int z = connectedSet.getInteger("pz"+i);
+			this.connectedSet.add(new BlockPos(x, y, z));
+		}
+
 		NBTTagCompound sendNodes = tag.getCompoundTag("SN");
 		for(String s : sendNodes.getKeySet()){
 			NodeSystem sys = new NodeSystem(this);
@@ -177,26 +236,17 @@ public abstract class Control {
 			receiveNodeMap.put(s, sys);
 			sys.readFromNBT(receiveNodes.getCompoundTag(s));
 		}
-		
-		NBTTagCompound customVarNames = tag.getCompoundTag("customvars");
-		for(int i = 0; i < customVarNames.getKeySet().size(); i ++){
-			this.customVarNames.add(customVarNames.getString("var"+i));
-		}
-		
-		NBTTagCompound connectedSet = tag.getCompoundTag("conset");
-		for(int i = 0; i < connectedSet.getKeySet().size()/3; i ++){
-			int x = connectedSet.getInteger("px"+i);
-			int y = connectedSet.getInteger("py"+i);
-			int z = connectedSet.getInteger("pz"+i);
-			this.connectedSet.add(new BlockPos(x, y, z));
-		}
-		
+
 		this.posX = tag.getFloat("X");
 		this.posY = tag.getFloat("Y");
 
 		NBTTagCompound configs = tag.getCompoundTag("configs");
 		for (String e : configs.getKeySet()) {
-			configMap.put(e, DataValue.newFromNBT(configs.getTag(e)));
+			DataValue value = DataValue.newFromNBT(configs.getTag(e));
+			if(value == null) {
+				throw new IllegalStateException("Failed to deserialize config '" + e + "' for control '" + registryName + "'");
+			}
+			configMap.put(e, value);
 		}
 	}