+++ src/main/java/com/hbm/inventory/control_panel/ControlEvent.java
@@ -1,17 +1,19 @@
 package com.hbm.inventory.control_panel;
 
+import it.unimi.dsi.fastutil.objects.Object2ObjectLinkedOpenHashMap;
+import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
 import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.util.EnumFacing;
 
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 
 public class ControlEvent {
 
-	private static final HashMap<String, ControlEvent> REGISTRY = new HashMap<>();
+	private static final Map<String, ControlEvent> REGISTRY = new Object2ObjectOpenHashMap<>();
 	
 	public String name;
-	public Map<String, DataValue> vars = new HashMap<>();
+	public Map<String, DataValue> vars = new Object2ObjectLinkedOpenHashMap<>();
 	
 	public ControlEvent(String name){
 		this.name = name;
@@ -46,7 +48,7 @@ public class ControlEvent {
 		ControlEvent evt = new ControlEvent(name);
 		//Set default values
 		for(Entry<String, DataValue> def : vars.entrySet()){
-			evt.vars.put(def.getKey(), def.getValue());
+			evt.vars.put(def.getKey(), def.getValue().copy());
 		}
 		return evt;
 	}
@@ -102,6 +104,14 @@ public class ControlEvent {
 		register(new ControlEvent("rbmk_crane_move").setVar("up", 0).setVar("down", 0).setVar("left", 0).setVar("right", 0));
 		register(new ControlEvent("rbmk_crane_load"));
 		register(new ControlEvent("ctrl_press").setVar("isSneaking", false));
+		register(new ControlEvent("redstone_input")
+				.setVar("facing", EnumFacing.NORTH)
+				.setVar("power", 0)
+				.setVar("weak", 0)
+				.setVar("strong", 0)
+				.setVar("isPowered", false)
+				.setVar("isWeaklyPowered", false)
+				.setVar("isStronglyPowered", false));
 		register(new ControlEvent("initialize"));
 	}
 }